Fixed certificate validation handling
Fix 1: the default for misp_ca_full_path is only used if the key doesn't exists. Once the setting is there, `None` is returned if the field is empty. Therefore, the enablement path is always chosen at this point.

Fix 2: the certificate validation is enabled independend from the verification checkbox setting. This is counterintuitive in the case that I want to keep the path and only disable for a short test.