feat: Add new mispgetattribute command
Adds the python script and modifies the config.

This command is useful if you want to fetch a single attribute
(/attributes/view/<id>) with a streaming command.
This allows you for example to verify that a single attribute is still
valid instead of having to check all of them. In case that an attribute
is hard deleted and not set to deleted=true properly, you can use this
command to figure that out.

The other commands have similar functionality, but are non streaming
commands.